feat: moved types to lib/types
