+++ apps/web/components/sections/project/teams/advisory-board.tsx
@@ -1,7 +1,7 @@
 import { Users } from "lucide-react";
 import { useState } from "react";
 import { Card } from "~/components/base/card";
-import type { ITeamMember } from "~/components/types/team";
+import type { ITeamMember } from "~/lib/types/project/team";
 
 export function AdvisoryBoardCard({ name, role, shortBio }: ITeamMember) { 
   return (
