+++ apps/web/components/sections/project/teams/leadership-team.tsx
@@ -1,6 +1,6 @@
 import { FaExternalLinkAlt, FaLink } from "react-icons/fa";
 import { Card } from "~/components/base/card";
-import type { ITeamMember } from "~/components/types/team";
+import type { ITeamMember } from "~/lib/types/project/team";
 import type { ITeamData } from "./advisory-board";
 
 export function LeadershipTeamCard({
